actualizo tp

+++ Tp1Lab/tp_laboratorio_1/funciones.h
@@ -9,7 +9,7 @@ int RestarNumeros(int,int);
 int MultiplicarNumeros(int,int);
 
 
-float DividirNumeros(float,int);
+float DividirNumeros(float,float);
 
 
 long int FactorizarNumero(int);
